# movies-ui/src/app/services/models/movie.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Rating } from '../models/rating';
export interface Movie {
  director?: string;
  genre?: string;
  id?: number;
  imdbID?: string;
  plot?: string;
  poster?: string;
  ratings?: Array<Rating>;
  title?: string;
  year?: string;
}